% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightControl.R
\name{add.knots}
\alias{add.knots}
\title{add linear interpolated points to flight tracks}
\usage{
add.knots(flights, res)
}
\arguments{
\item{flights}{the flights data.frame}

\item{res}{spatial resolution}
}
\value{
a new flights data.frame with more rows, but some neglected columns
}
\description{
add linear interpolated points to flight tracks
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

