% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightControl.R
\name{airplanes}
\alias{airplanes}
\title{read the air traffic csv file}
\usage{
airplanes(file, raw = FALSE, max.altitude = 5000, unit = "m")
}
\arguments{
\item{file}{path to and name of the csv file}

\item{raw}{logical i ftrue returns the full data set otherwise extract only position/altitude/time with callSign (flightID)}

\item{max.altitude}{ignore positions above that altitude in the given unit}

\item{unit}{default m(eter). If anything else than f(eet) is given the default is used}
}
\value{
data.frame
}
\description{
reads a csv file recorded by \href{https://github.com/antirez/dump1090}{dump1090} and returns it as data.frame
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}
\references{
\href{https://github.com/antirez/dump1090}{dump1090}
}

