% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightControl.R
\name{animate.flights.2d}
\alias{animate.flights.2d}
\title{2D map animation of series of pictures}
\usage{
animate.flights.2d(flights, step = 5)
}
\arguments{
\item{flights}{the flights data.frame}

\item{step}{a new picture every "step" minutes"}
}
\description{
Creates a time series of images in tempdir(). If present it uses ffmpeg to create an animation out of them.
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

