% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightControl.R
\name{animate.flights.3d}
\alias{animate.flights.3d}
\title{3D isosurface of lowest flight altitude}
\usage{
animate.flights.3d(flights, dem, tracks = TRUE)
}
\arguments{
\item{flights}{the flights data.frame}

\item{dem}{digital elevation model}

\item{tracks}{logical if the individual tracks should be shown (default) or a transparent 'carpet' of the spatially interpolated lowest flight altitude.}
}
\description{
3D isosurface of lowest flight altitude
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

