% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{breakCallSign}
\alias{breakCallSign}
\title{rename the callSign at a temporal break}
\usage{
breakCallSign(flights, pause.limit = 900)
}
\arguments{
\item{flights}{the flights data.frame}

\item{pause.limit}{which temporal delta time should be considered as new flight}
}
\value{
new flights data.frame
}
\description{
A random string of six characters is attached to the callSign each time it appears again after more than `pause.limit` seconds.
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

