% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightControl.R
\name{convert}
\alias{convert}
\title{Convert csv to RData}
\usage{
convert(dates, source.dir = ".", target.dir = ".", verbose = TRUE, ...)
}
\arguments{
\item{dates}{a vector of dates}

\item{source.dir}{path to where the flights_<yyyymmdd>.csv files are}

\item{target.dir}{path, where to save the RData files}

\item{verbose}{print some information}

\item{...}{additional parameters passed on to \code{\link{airplanes}}}
}
\description{
Convert daily csv files produced by dump1080 to RData files
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

