% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter.position}
\alias{filter.position}
\title{Filter for a certain region}
\usage{
filter.position(flights, west = -180, east = 180, south = -90,
  north = 90, length = 1)
}
\arguments{
\item{flights}{the flights data.frame}

\item{west}{western edge}

\item{east}{eastern edge}

\item{south}{southern edge}

\item{north}{northern edge}

\item{length}{minimum number of nodes per flight track}
}
\value{
new flights data.frame
}
\description{
Applies some subsets by coordinates
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

