% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flights}
\alias{flights}
\title{flights}
\description{
This dataset contains unsplit flight tracks of October 28th and 30th, 2016 below 3000 meters.
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}
\keyword{data}

