% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{flightsDirectionSplit}
\alias{flightsDirectionSplit}
\title{Applies a filter by total altitudinal change}
\usage{
flightsDirectionSplit(flights, direction = NA, threshold = 500)
}
\arguments{
\item{flights}{the flights data.frame}

\item{direction}{character, either"a(scending)", "d(escending)" or "p(assing)". For anything else the full flights data.frame is returned including a new column `dz`.}

\item{threshold}{absolute treshold value to distinguish between the the previous classes.}
}
\value{
new flights data.frame
}
\description{
Creates subsets, if the plane is descending, ascending or passing by, based on a threshold value of total altitude change
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

