% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{import}
\alias{import}
\title{get the flights table in the workspace}
\usage{
import(dates, source.dir = ".", verbose = TRUE)
}
\arguments{
\item{dates}{list of dates}

\item{source.dir}{path to where the RData flights_<yyyymmdd>.RData files are}

\item{verbose}{print some information during processing}
}
\value{
a data.frame of all flights
}
\description{
Loads one or more RData files produced by \code{\link{convert}} into one data.frame
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

