% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightControl.R
\name{min.flight.altitude}
\alias{min.flight.altitude}
\title{calculate a surface of the minimum flight height}
\usage{
\method{min}{flight.altitude}(flights, extent = NULL, res = NULL,
  proj = NULL, method = "idw")
}
\arguments{
\item{flights}{flights data.frame as returned by}

\item{extent}{a spatial extent object}

\item{res}{target resolution}

\item{proj}{target projection}

\item{method}{interpolation method (see Details)}
}
\value{
a interpolated SpatialPixelsDataFrame of minimum flight altitude
}
\description{
Interpolation methods:
\itemize{
\item{raw}{no interpolation}
\item{idw}{kriging see \code{\link[gstst]{krige}}}
\item{sph/gau}{variogram model see \code{\link[gstat]{vgm}}}
}
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

