% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{smoothFlightsGrid}
\alias{smoothFlightsGrid}
\title{Spatial smoothing}
\usage{
smoothFlightsGrid(data, method = "loess", k = 5, ...)
}
\arguments{
\item{data}{a SpatialPixelsDataFrame}

\item{method}{interpolation method. One of 'loess' (default), 'gam', 'gau' or 'sph'}

\item{k}{if method='gam' parameter for the smooth term in te}

\item{...}{further parameters passed on to 'loess'.}
}
\value{
a SpatialPixelsDataFrame
}
\description{
An object produced by \code{\link{gridFlights}} is spatially interpolated to the whole rectangular extent. Four interplolation methods are implemented:
\enumerate{
\item \code{\link{loess}} smoothest surface, not close to reality
\item \code{\link{gam}} smooth surface, closer to reality. For higher `k` values closer to reality, but computation time increases, too.
\item \code{\link[gstat]{krige}} 'Sph'
\item \code{\link[gstat]{krige}} 'Gau'
}
}
\examples{
require(raster)
extent <- extent(c(7.25, 8.5, 49.5, 50.2))
N <- 750
set.seed(987654)
data <- data.frame(lon=rbeta(N, 5, 2) * (extent@xmax - extent@xmin) + extent@xmin,
                   lat=rbeta(N, 2.5, 2.5) * (extent@ymax - extent@ymin) + extent@ymin)
data$altitude = (data$lon - 7.9)^3 - 3 * (data$lon - 7.9) * (data$lat- 50)^2
fgrid = gridFlights(data, min)
res <- smoothFlightsGrid(fgrid, "loess")
names(res) <- "loess"
res2 <- smoothFlightsGrid(fgrid, "gam")
res@data$gam = res2@data$altitude
res2 <- smoothFlightsGrid(fgrid, "gau")
res@data$gau = res2@data$altitude
res2 <- smoothFlightsGrid(fgrid, "sph")
res@data$sph = res2@data$altitude
spplot(res)
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

