% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{split.pause.callSign}
\alias{split.pause.callSign}
\title{split the callSign by a temporal break}
\usage{
split.pause.callSign(flights, pause.limit = 900)
}
\arguments{
\item{flights}{the flights data.frame}

\item{pause.limit}{which temporal delta time should be considered as new flight}
}
\value{
new flights data.frame
}
\description{
split the callSign by a temporal break
}
\author{
Joerg Steinkamp \email{joergsteinkamp@yahoo.de}
}

